Separate utilities from procedures in `child-lab-procedures`
commit-id:dd413fcf
